Move the unit test to unittest_brain because it needs six
